Move addfactorsource package
