+++ app/src/main/java/com/babylon/wallet/android/di/UiModule.kt
@@ -3,9 +3,9 @@ package com.babylon.wallet.android.di
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesIOHandler
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesProxy
 import com.babylon.wallet.android.presentation.accessfactorsources.AccessFactorSourcesProxyImpl
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.AddFactorSourceIOHandler
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.AddFactorSourceProxy
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.AddFactorSourceProxyImpl
+import com.babylon.wallet.android.presentation.addfactorsource.AddFactorSourceIOHandler
+import com.babylon.wallet.android.presentation.addfactorsource.AddFactorSourceProxy
+import com.babylon.wallet.android.presentation.addfactorsource.AddFactorSourceProxyImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
