+++ app/src/main/java/com/babylon/wallet/android/presentation/addfactorsource/AddFactorSourceNav.kt
@@ -1,14 +1,14 @@
-package com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource
+package com.babylon.wallet.android.presentation.addfactorsource
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptionsBuilder
 import androidx.navigation.navigation
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.device.confirmseedphrase.confirmDeviceSeedPhrase
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.device.seedphrase.deviceSeedPhrase
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.intro.ROUTE_ADD_FACTOR_INTRO
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.intro.addFactorIntro
-import com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.name.setFactorName
+import com.babylon.wallet.android.presentation.addfactorsource.device.confirmseedphrase.confirmDeviceSeedPhrase
+import com.babylon.wallet.android.presentation.addfactorsource.device.seedphrase.deviceSeedPhrase
+import com.babylon.wallet.android.presentation.addfactorsource.intro.ROUTE_ADD_FACTOR_INTRO
+import com.babylon.wallet.android.presentation.addfactorsource.intro.addFactorIntro
+import com.babylon.wallet.android.presentation.addfactorsource.name.setFactorName
 
 const val ROUTE_ADD_FACTOR_SOURCE_GRAPH = "add_factor_source_graph"
