+++ app/src/main/java/com/babylon/wallet/android/presentation/addfactorsource/AddFactorSourceProxyContract.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource
+package com.babylon.wallet.android.presentation.addfactorsource
 
 import com.radixdlt.sargon.FactorSourceId
 import com.radixdlt.sargon.FactorSourceKind
