+++ app/src/main/java/com/babylon/wallet/android/presentation/addfactorsource/AddFactorSourceProxyImpl.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource
+package com.babylon.wallet.android.presentation.addfactorsource
 
 import com.babylon.wallet.android.utils.AppEvent
 import com.babylon.wallet.android.utils.AppEventBus
