+++ app/src/main/java/com/babylon/wallet/android/presentation/addfactorsource/device/confirmseedphrase/ConfirmDeviceSeedPhraseViewModel.kt
@@ -1,4 +1,4 @@
-package com.babylon.wallet.android.presentation.settings.securitycenter.addfactorsource.device.confirmseedphrase
+package com.babylon.wallet.android.presentation.addfactorsource.device.confirmseedphrase
 
 import androidx.lifecycle.viewModelScope
 import com.babylon.wallet.android.data.repository.factors.DeviceMnemonicBuilderClient
